Improve test names, rename file
