+++ lib/liquid/autoescape/liquid_ext/block_body.rb
