+++ lib/liquid/autoescape/tags/autoescape.rb
@@ -1,5 +1,5 @@
 require "liquid"
-require "liquid/autoescape/liquid_ext/block"
+require "liquid/autoescape/liquid_ext/block_body"
 require "liquid/autoescape/liquid_ext/capture"
 require "liquid/autoescape/liquid_ext/standard_filters"
